apps/api/pyproject.toml
[tool.poetry]
name = "ai-research-api"
version = "0.1.0"
description = "FastAPI backend for AI Research Assistant"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pydantic = "^2.0.0"
openai = "^1.12.0"
pypdf = "^3.17.0"
langchain = "^0.0.340"
pinecone-client = "^2.2.4"
arxiv = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
isort = "^5.10.1"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
